% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmETM.R
\name{genETMTree}
\alias{genETMTree}
\title{Generation of rooted binary trees under the equiprobable-types-model (ETM)}
\usage{
genETMTree(n)
}
\arguments{
\item{n}{Integer value that specifies the desired number of leaves, i.e.,
vertices with in-degree 1 and out-degree 0.\cr
Due to the restrictions of the \code{phylo} or \code{multiphylo} format,
the number of leaves must be at least 2 since there must be at
least one edge.}
}
\value{
\code{genETMTree} A single tree of class \code{phylo} is returned.
}
\description{
\code{genETMTree} - Generates a rooted binary tree in \code{phylo}
format with the given number of \code{n} leaves under the
equiprobable-types-model.
Given \code{n}, all tree shapes/topologies with \code{n} leaves are
equiprobable under the ETM.
}
\examples{
genETMTree(n = 5)
}
\references{
\itemize{
\item This function uses the \code{rtree(..., equiprob = T)} function of the
\code{ape} package
(E. Paradis, K. Schliep. “ape 5.0: an environment for modern
phylogenetics and evolutionary analyses in R.”
Bioinformatics, 35, 526-528, 2019).
}
}
