% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmPDA.R
\name{genPDATree}
\alias{genPDATree}
\title{Generation of rooted binary trees under the PDA model}
\usage{
genPDATree(n)
}
\arguments{
\item{n}{Integer value that specifies the desired number of leaves, i.e.,
vertices with in-degree 1 and out-degree 0.}
}
\value{
\code{genPDATree} A single tree of class \code{phylo} is returned.
}
\description{
\code{genPDATree} - Generates a rooted binary tree in \code{phylo}
format with the given number of \code{n} leaves under the
proportional-to-distinguishable-arrangements model.
Given \code{n}, all phylogenies (trees with labeled leaves) with
\code{n} leaves are equiprobable under the PDA.\cr
Due to the restrictions of the \code{phylo} or \code{multiphylo} format
the number of leaves must be at least 2 since there must be at
least one edge.
}
\examples{
genPDATree(n = 5)
}
\references{
\itemize{
\item This function uses the \code{rtopology(..., rooted = T)} function of
the \code{ape} package
(E. Paradis, K. Schliep. “ape 5.0: an environment for modern
phylogenetics and evolutionary analyses in R.”
Bioinformatics, 35, 526-528, 2019). \cr
\item D. E. Rosen. Vicariant patterns and historical explanation in
biogeography. Systematic Zoology, 27(2):159, 1978.
}
}
\author{
Sophie Kersting
}
